requirements.txt
fastapi
uvicorn
python-docx
PyMuPDF
scikit-learn
joblib
requests
beautifulsoup4
